Add bulk tests

+++ Domain/Models/VwElectricityDatum.cs
@@ -4,7 +4,7 @@
     {
         public string? Pavadinimas { get; set; }
         public string? RegionName { get; set; }
-        public long? Numeris { get; set; }
+        public int? Numeris { get; set; }
         public decimal? Ppliusas { get; set; }
         public DateTime? Plt { get; set; }
         public decimal? Pminusas { get; set; }
